pyproject.toml
[tool.poetry]
name = "musicalstore"
version = "0.1.0"
description = ""
authors = ["Roman <roma.makhunov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.3"
pillow = "^10.2.0"
python-slugify = "^8.0.4"
django-mptt = "^0.16.0"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.3.0"
faker = "^24.4.0"
requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
